Renaming tensorflow.py
